Reoplacing json config file with enviorment vars
We used to read configuration constants from defined json objects in the /config directory. This solution was problematic for the deplyoment with heroku and docker

Now we read all enviorment variables directly from the .env file in the project.

A helper function has been added to generate the .env file with the required enviorment variabels for a service.
The setup function reads a .sample-env file, witch contains all nessecary enviorment variables that are going to be used.