Configuration changes to apply validation defintions
Earlier we used to create default data when migrating tables with the library knex.

With this change we have now take advantage of  knex seed functionallity to populate the database with default data.
Since validations is not the only default data we use this commit also includes the movement of eralier defined default data.