Feature/api version numbers (#16)
* Support for json-api-serializer (#10)

Implemented json-api-serilizer with additional helper function to convert KNEX queries into seraializable data objects.

* Simple versioning added

Cahnged the base route for the service to "api/v1" and added correct version response based on the conflunce documentation. This change does not include listing of available versions for the api on the root route "/".