Feature/options data (#20)
* Support for json-api-serializer (#10)

Implemented json-api-serilizer with additional helper function to convert KNEX queries into seraializable data objects.

* Serializing options data for a question to jsonapi

The retriving of options for a question was not working as expected. This problem was caused due to a foreigin_key contraint not beeing set correctly and also a naming issue in bookshelf. The problem caused no options to be returned when requesting a question with options.

Fixed the syntax when then foreigin key contraint was defined. On top of this the options key in the data that is created from a knex query was changed from choices to options, the name changed made the data included in the jsonapi serilization that is returned in the response.

* Removed console.logs