Structure for retrival of questions with logic dependecy
Upon requesting a collection of questions a single question now includes the key logics. The logics key has data for handle dependecies to later questions in a form.

The logic object that is recived contains the properties logic_type, actions and trigger.

The logic_type and trigger identify the triggers for the actions.

The actions is an array of objects that defines the logic's behavior.